Fix security hole involving malicious directory renames.
Before the fix malicious user could either create "$TMPLINKDIR" (our
'mkdir -p' wouldn't fail in this case) or she/he could create
"$TMPLINKDIR"/libkroki-glxoffload.so.0 as a directory thus in both cases
gain the control over the link to the library that we are going to execute.

Thanks to Alexander Monakov for raising the concern.