Fix issue #1: undefined symbol: glPixelStorei
By default ld creates objects with now obsolete DT_RPATH that is searched
_before_ LD_LIBRARY_PATH.  Turned out I never tried to actually install
kroki/glxoffload but did only `make DESTDIR=/tmp'.  With this DT_RPATH for
libkroki-glxoffload.so lookup was (erroneously?) pointing to
/usr/local/lib/kroki-glxoffload/stub (default destdir), no library was
found there and LD_LIBRARY_PATH was searched as intended.  However after
installing kroki/glxoffload the stub library was found first (it should
be noted that it is impossible to not install the stub library as this
would either break linking with it or RPM dependency on it).

With ld --enable-new-dtags DT_RPATH records are ignored and new DT_RUNPATH
records are searched _after_ LD_LIBRARY_PATH, whish is what has been
intended from the start.